requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
django-dirtyfields==1.3.1
djangorestframework==3.9.2
pytz==2018.9
